fix responseFunctions2 state filtering
The previous algorithm to find states that could be accessed used
powers of the creation operator to climb the ladder. With states with a
bilinear coupling, this algorithm could miss needed states due to a
cancellation. (They looked dark even when they weren’t.) For example,
two carbonyls that are not parallel gave only the antisymmetric band
and no symmetric band (even though it should). The new algorithm uses
the transition dipole operators and searches for the dipole to be above
some cutoff. This seems to have fixed the issue.