Added insertion of entry template upon creation of new LA page
When the labarchivesCallObj creates a new page it will automatically insert a new entry template into the new notebook page. This was previously being done in by the FileSystem class in the Mahou repo but was inserting an entry template every time a FileSystem object was initialized instead of only when the page was created.