Working on incorporating WiaC model to our lineshape functions for simulations.
Created many test lineshape functions that use numerical integration (NI) of g(t) for a single exponential decay model, single wobbling cone model, and a single diffusive model. We also changed the lineshapeFunctionBnd to remove complications with defining aRFoption.order for the wobbling in a cone model. It does not change outputs in other lineshape functions.